rename: 스탯 dto 패키지 변경
