+++ src/main/java/com/ddudu/application/domain/ddudu/service/DduduDomainService.java
@@ -3,9 +3,9 @@ package com.ddudu.application.domain.ddudu.service;
 import com.ddudu.application.annotation.DomainService;
 import com.ddudu.application.domain.ddudu.domain.Ddudu;
 import com.ddudu.application.domain.user.domain.User;
-import com.ddudu.application.dto.ddudu.StatsBaseDto;
 import com.ddudu.application.dto.ddudu.request.CreateDduduRequest;
-import com.ddudu.application.dto.goal.MonthlyStatsSummaryDto;
+import com.ddudu.application.dto.stats.MonthlyStatsSummaryDto;
+import com.ddudu.application.dto.stats.StatsBaseDto;
 import java.time.LocalDate;
 import java.time.YearMonth;
 import java.util.List;
