+++ src/main/java/com/ddudu/application/port/in/ddudu/CollectMonthlyStatsSummaryUseCase.java
@@ -1,6 +1,6 @@
 package com.ddudu.application.port.in.ddudu;
 
-import com.ddudu.application.dto.goal.response.MonthlyStatsSummaryResponse;
+import com.ddudu.application.dto.stats.response.MonthlyStatsSummaryResponse;
 import java.time.YearMonth;
 
 public interface CollectMonthlyStatsSummaryUseCase {
