+++ src/main/java/com/ddudu/application/service/ddudu/CollectMonthlyStatsSummaryService.java
@@ -4,9 +4,9 @@ import com.ddudu.application.annotation.UseCase;
 import com.ddudu.application.domain.ddudu.exception.DduduErrorCode;
 import com.ddudu.application.domain.ddudu.service.DduduDomainService;
 import com.ddudu.application.domain.user.domain.User;
-import com.ddudu.application.dto.ddudu.StatsBaseDto;
-import com.ddudu.application.dto.goal.MonthlyStatsSummaryDto;
-import com.ddudu.application.dto.goal.response.MonthlyStatsSummaryResponse;
+import com.ddudu.application.dto.stats.MonthlyStatsSummaryDto;
+import com.ddudu.application.dto.stats.StatsBaseDto;
+import com.ddudu.application.dto.stats.response.MonthlyStatsSummaryResponse;
 import com.ddudu.application.port.in.ddudu.CollectMonthlyStatsSummaryUseCase;
 import com.ddudu.application.port.out.goal.MonthlyStatsPort;
 import com.ddudu.application.port.out.user.UserLoaderPort;
