+++ src/test/java/com/ddudu/application/domain/ddudu/service/DduduDomainServiceTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.ddudu.application.domain.ddudu.domain.Ddudu;
 import com.ddudu.application.domain.ddudu.domain.enums.DduduStatus;
 import com.ddudu.application.domain.user.domain.User;
-import com.ddudu.application.dto.ddudu.StatsBaseDto;
 import com.ddudu.application.dto.ddudu.request.CreateDduduRequest;
+import com.ddudu.application.dto.stats.StatsBaseDto;
 import com.ddudu.fixture.DduduFixture;
 import com.ddudu.fixture.GoalFixture;
 import com.ddudu.fixture.UserFixture;
